rename file, fix path
